Confirmed server properly receives correct data from database
